backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
numpy==2.2.0
faiss-cpu==1.9.0.post1
google-generativeai==0.8.3
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.0.0
pyserial==3.5
requests==2.32.3
python-dotenv==1.0.1
transformers==4.48.2
torch==2.6.0
torchaudio==2.6.0
elevenlabs==1.13.4
librosa==0.10.2.post1
soundfile==0.13.1
huggingface_hub==0.27.1
